requirements.txt
line-bot-sdk
Flask
requests
python-dotenv
pymongo
openai